-updated constructor to send 10 tickets of each type to owner -began writing tests -encountered "Transaction reverted: function selector was not recognized and there's no fallback function" error when calling startRaffle() w/ items

+++ contracts/RaffleContract.sol
@@ -19,7 +19,7 @@ struct Raffle {
     mapping(address => UserStake[]) userStakes;
     mapping(address => bool) prizeClaimed;
     uint256 randomNumber;
-    uint32 raffleEnd;
+    uint256 raffleEnd;
 }
 
 struct UserStake {
@@ -75,7 +75,7 @@ contract RaffleContract {
         require(msg.sender == s.contractOwner, "Raffle: Must be contract owner");
         emit RaffleStarted(s.raffles.length, _raffleEnd, _raffleItems);
         Raffle storage raffle = s.raffles.push();
-        raffle.raffleEnd = uint32(_raffleEnd);
+        raffle.raffleEnd = uint256(_raffleEnd);
         for (uint256 i; i < _raffleItems.length; i++) {
             address stakeAddress = _raffleItems[i].stakeAddress;
             uint256 stakeId = _raffleItems[i].stakeId;
